be/int/web/registration: moves up the registration agent, splitting from receptionist and middlewares.
