+++ backend/cmd/objectives/main.go
@@ -12,7 +12,7 @@ import (
 	sessions "logbook/cmd/sessions/client"
 	"logbook/internal/startup"
 	"logbook/internal/web/balancer"
-	"logbook/internal/web/registration/reception"
+	"logbook/internal/web/registration"
 	"logbook/internal/web/registryfile"
 	"logbook/internal/web/router"
 	"logbook/internal/web/sidecar"
@@ -48,7 +48,7 @@ func Main() error {
 	pub := endpoints.NewPublic(a, sessions.NewClient(balancer.New(sc.InstanceSource(models.Sessions))), l)
 	priv := endpoints.NewPrivate(a, l)
 
-	agent := reception.NewAgent(deplcfg, l)
+	agent := registration.NewAgent(deplcfg, l)
 	err = agent.RegisterEndpoints(pub, priv)
 	if err != nil {
 		return fmt.Errorf("agent.RegisterEndpoints: %w", err)
